requirements.txt
fastapi
web3
uvicorn
pyjwt
python-dotenv
pydantic[email]
